preexec within BaseCmd.__call__; postexec within Sequnce.go
While preexec should stall the execution of its own cmd, it is called
within BaseCmd.__call__ to not also stall the execution of following
cmds when running as non-joined Thread.

But postexec only makes sense when stalling the following cmds, so it is
called within Sequence.go. This way followers will be stalled even for
non-joined Threads.

Besides Sequence.add_cmd add itself to the cmd, so it has not to be
passed to BaseCmd.check and other methods.
It is always available as BaseCmd-obj.sequence.